# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kasir API",
    "description": "A simple POS (Point of Sale) REST API built with Go, implementing clean Layered Architecture pattern.",
    "version": "1.0.0",
    "contact": {
      "name": "Kasir API"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://kasir-api-production-ecd5.up.railway.app",
      "description": "Production server (Railway)"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["General"],
        "summary": "API Information",
        "description": "Returns API information and list of available endpoints",
        "responses": {
          "200": {
            "description": "API information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiInfo"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["General"],
        "summary": "Health Check",
        "description": "Check if the server is running and healthy",
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/produk": {
      "get": {
        "tags": ["Products"],
        "summary": "Get All Products",
        "description": "Retrieve all products. Supports optional search by product name.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "description": "Search products by name (partial match)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create a Product",
        "description": "Create a new product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              },
              "example": {
                "name": "Nasi Goreng",
                "price": 15000,
                "stock": 100
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    },
    "/api/produk/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get Product by ID",
        "description": "Retrieve a single product by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product ID"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update Product",
        "description": "Update an existing product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              },
              "example": {
                "name": "Nasi Goreng Special",
                "price": 20000,
                "stock": 50
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or product ID"
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Delete Product",
        "description": "Delete a product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Product ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product ID"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get All Categories",
        "description": "Retrieve all categories",
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create a Category",
        "description": "Create a new category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              },
              "example": {
                "name": "Makanan",
                "description": "Kategori produk makanan"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get Category by ID",
        "description": "Retrieve a single category by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Category ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid category ID"
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "put": {
        "tags": ["Categories"],
        "summary": "Update Category",
        "description": "Update an existing category by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Category ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              },
              "example": {
                "name": "Minuman",
                "description": "Kategori produk minuman"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or category ID"
          }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete Category",
        "description": "Delete a category by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Category ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid category ID"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/checkout": {
      "post": {
        "tags": ["Transactions"],
        "summary": "Checkout / Create Transaction",
        "description": "Process a checkout with multiple items. Each item requires a product_id and quantity. Stock will be reduced automatically.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              },
              "example": {
                "items": [
                  { "product_id": 1, "quantity": 2 },
                  { "product_id": 2, "quantity": 1 }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "500": {
            "description": "Internal server error (e.g. insufficient stock)"
          }
        }
      }
    },
    "/api/report/hari-ini": {
      "get": {
        "tags": ["Reports"],
        "summary": "Today's Sales Report",
        "description": "Get sales summary for today including total revenue, total transactions, and best selling product",
        "responses": {
          "200": {
            "description": "Today's sales summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesSummary"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/report": {
      "get": {
        "tags": ["Reports"],
        "summary": "Sales Report by Date Range",
        "description": "Get sales summary for a specific date range",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "description": "Start date (format: YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-01-01"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "description": "End date (format: YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2026-02-01"
          }
        ],
        "responses": {
          "200": {
            "description": "Sales summary for the specified date range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesSummary"
                }
              }
            }
          },
          "400": {
            "description": "start_date and end_date are required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiInfo": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Selamat datang di Kasir API"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "endpoints": {
            "type": "object"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "example": "Server is healthy"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Nasi Goreng"
          },
          "price": {
            "type": "integer",
            "example": 15000
          },
          "stock": {
            "type": "integer",
            "example": 100
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "required": ["name", "price", "stock"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Nasi Goreng"
          },
          "price": {
            "type": "integer",
            "example": 15000
          },
          "stock": {
            "type": "integer",
            "example": 100
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Makanan"
          },
          "description": {
            "type": "string",
            "example": "Kategori produk makanan"
          }
        }
      },
      "CategoryInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Makanan"
          },
          "description": {
            "type": "string",
            "example": "Kategori produk makanan"
          }
        }
      },
      "CheckoutRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckoutItem"
            }
          }
        }
      },
      "CheckoutItem": {
        "type": "object",
        "required": ["product_id", "quantity"],
        "properties": {
          "product_id": {
            "type": "integer",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "example": 2
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "total_amount": {
            "type": "integer",
            "example": 45000
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-08T10:30:00Z"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionDetail"
            }
          }
        }
      },
      "TransactionDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "transaction_id": {
            "type": "integer",
            "example": 1
          },
          "product_id": {
            "type": "integer",
            "example": 1
          },
          "product_name": {
            "type": "string",
            "example": "Nasi Goreng"
          },
          "quantity": {
            "type": "integer",
            "example": 2
          },
          "subtotal": {
            "type": "integer",
            "example": 30000
          }
        }
      },
      "SalesSummary": {
        "type": "object",
        "properties": {
          "total_revenue": {
            "type": "integer",
            "example": 150000
          },
          "total_transaksi": {
            "type": "integer",
            "example": 5
          },
          "produk_terlaris": {
            "$ref": "#/components/schemas/BestSeller"
          }
        }
      },
      "BestSeller": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Nasi Goreng"
          },
          "qty_terjual": {
            "type": "integer",
            "example": 20
          }
        }
      },
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Product deleted successfully"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "General",
      "description": "General endpoints"
    },
    {
      "name": "Products",
      "description": "Product management (CRUD)"
    },
    {
      "name": "Categories",
      "description": "Category management (CRUD)"
    },
    {
      "name": "Transactions",
      "description": "Checkout and transaction processing"
    },
    {
      "name": "Reports",
      "description": "Sales reports and analytics"
    }
  ]
}